fix up scrubbing tests
